fix bordersize set to 0 with no ini file available
Previously, the ability to set the border size to 0 would only
happen after the border size had been initialized to 8, which
properly initialized the window size and text position.  So if you
changed the border size, an ini file would be created with properly
calculated window sizes and text positions.  Without that, the fixup
code would fail due to an edge case.  I just fixed up the logic and it
works now.